// src/types/patient.ts
export type WeightAccuracyValue = 'direct' | 'estimate'
export type WeightEstimateByValue = 'by-age' | 'by-height'
export type SexValue = 'male' | 'female'
export type HabitusModeValue = 'child' | 'adult'
export type HabitusValue =
  | 'very-thin'
  | 'thin'
  | 'normal'
  | 'sporty'
  | 'lightly-overweight'
  | 'overweight'
  | 'very-overweight'

export type ResultTypeValue =
  | 'direct'
  | 'by-age'
  | 'by-height-curve'
  | 'by-height-bmi'
  | 'undefined'